Client/Server: Update - home route, fetch lib,  updated

+++ src/app/api/broker-accounts/route.ts
@@ -1,5 +1,10 @@
 import connectMongoDB from "@/libs/mongodb";
 import BankAccounts from "@/models/bank-account";
+import Bond from "@/models/bond";
+import FundB from "@/models/fundB";
+import FundS from "@/models/fundS";
+import Stock from "@/models/stock";
+
 import { NextResponse, NextRequest } from "next/server";
 
 
@@ -50,6 +55,20 @@ export async function DELETE(request: NextRequest) {
   const id = request.nextUrl.searchParams.get("id");
 
   await connectMongoDB();
+
+  const bonds = await Bond.find({ brokerId: id });
+  const stocks = await Stock.find({ brokerId: id });
+  const funds = await FundS.find({ brokerId: id });
+  const fundsB = await FundB.find({ brokerId: id });
+
+
+
+  if (bonds.length > 0 || stocks.length > 0 || funds.length > 0 || fundsB.length > 0) {
+    return NextResponse.json(
+      { message: "Broker account cannot be deleted because it has assets" },
+      { status: 400 }
+    )
+  }
 
   await BankAccounts.findByIdAndDelete(id);
